+++ src/Network/Network.Test/ScenarioTests/AzureFirewallPolicyTests.cs
@@ -212,5 +212,21 @@ namespace Commands.Network.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-AzureFirewallPolicyIDPSProfiles");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.azurefirewall)]
+        public void TestAzureFirewallPolicyDraft()
+        {
+            TestRunner.RunTestScript("Test-AzureFirewallPolicyDraft");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.azurefirewall)]
+        public void TestAzureFirewallPolicyRCGyDraft()
+        {
+            TestRunner.RunTestScript("Test-AzureFirewallPolicyRCGDraft");
+        }
     }
 }